UserManagement/api/requirements.txt
# requirements.txt

argon2==0.1.10 #for password hashing
asgiref==3.8.1 #for django
Django==5.0.6 #for backend
django-filter==24.2 #for filtering
djangorestframework==3.15.1 #for rest api
djangorestframework-simplejwt==5.3.1 #for jwt
pillow==10.3.0 #for image handling
PyJWT==2.8.0 #for jwt
python-dotenv==1.0.1 #for env variables load_dotenv
sqlparse==0.5.0 #for postgres
typing_extensions==4.12.0	# for python 3.6
psycopg[binary]==3.1.12 #postgres db
django-cors-headers==4.3.1 #cors to allow frontend to access backend
argon2-cffi==23.1.0 #password hashing
requests==2.31.0 #for making requests in oauth part
cryptography==42.0.7 #for encryption of state in oauth part
django-crontab==0.7.1 #for cron jobs
